// frontend/src/types/doctor.ts
export type UserDoctor = {
  id: number;
  email: string;
  role_id: number;
  nombre: string;
  apellido: string;
  numero_telefono?: string;
  direccion?: string;
  sexo?: string;
  password?: string;
  especialidad?: string;
};

export type UseDoctorsReturn = {
  doctors: UserDoctor[];
  newDoctor: {
    email: string;
    password: string;
    role_id: number;
    nombre: string;
    apellido: string;
    numero_telefono?: string;
    direccion?: string;
    sexo?: string;
    especialidad?: string;
  };
  editDoctor: UserDoctor | null;
  setNewDoctor: React.Dispatch<
    React.SetStateAction<{
      email: string;
      password: string;
      role_id: number;
      nombre: string;
      apellido: string;
      numero_telefono?: string;
      direccion?: string;
      sexo?: string;
      especialidad?: string;
    }>
  >;
  setEditDoctor: React.Dispatch<React.SetStateAction<UserDoctor | null>>;
  fetchDoctors: () => Promise<void>;
  handleCreateOrUpdate: (e: React.FormEvent) => Promise<void>;
  handleDelete: (id: number) => Promise<void>;
  ConfirmDialog: () => React.JSX.Element;
};

export type UserData = {
  id?: number;
  email: string;
  password?: string;
  nombre: string;
  apellido: string;
  numero_telefono?: string;
  direccion?: string;
  sexo?: string;
  role_id: number;
  especialidad?: string;
};